requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-experimental==0.0.50
pandas==2.2.0
plotly==5.18.0
seaborn==0.13.2
matplotlib==3.8.2
requests==2.31.0
pydantic==2.6.0
python-dotenv==1.0.1
